requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
PyMuPDF==1.25.1
requests==2.32.3
reportlab==4.2.5
python-dotenv==1.0.1
pydantic==2.10.4
